Adding new run-crons.sh script
This script can be used to run all cronjobs required for `mmcfe-ng`.
It can be expanded by pool users in the future by adding or removing
job from the CRONS list in the script.

Script assumed to run inside the cronjobs folder, this can be adjusted
via CRONHOME in the script.