[FIX] Coins without nethashrate in getmininginfo
Fixes issues with coins that did not implement a proper array index into
their `getmininginfo` output. Currently supported keys are

* networkhashps
* hashespersec
* netmhashps

If none of there are found or `getmininginfo` fails itself, we now
default to `0` to ensure the webfrontend still works.

Fixes #847 and also fixes #834 which has not reported back the
required information.