Fix occasional error: unable to unserialize data
Fix Memcache occasional error "unable to unserialize data" that occurs when running cronjobs.

In PHP documentation, the 3 parameter for set is a flag and the 4th parameter is the expiration.
http://www.php.net/manual/en/memcache.set.php

Making this update seems to eliminate the "...unable to unserialize data..." error.  Not sure why that solves it, but it did for me.