(#1717) Added BitcoinWrapper::validateaddress() function
Without this patch, the call to $bitcoin->validateaddress() in
public/include/admin_checks.php always fails, causing the admin to
be informed that their 'cold wallet address is SET and INVLAID'.
This happens as long as a cold wallet address is defined,
regardless of if the address is valid or not.

The problem can be traced back to
public/include/classes/bitcoin.class.php where we define the
BitcoinClient::validateaddress() function. This function simply
returns true or false depending on the result of a call to
parent::validateaddress($coin_address). In BitcoinClient's case,
parent is always BitcoinWrapper. However, we have never defined
BitcoinWrapper::validateaddress().

To resolve this, I simply defined the
BitcoinWrapper::validateaddress() function.