Fix bug in admin dashboard
Fixes a bug causing the admin dashboard to throw an error, when the wallet has an account name which is an integer (e.g. "1").

This is caused by json_decode decoding the account address as an integer, which is passed to getaddressesbyaccount, throwing an error (as the daemon only accepts string values for account names).