Remove MEGA Crap From Getting Started
What's the go with all the MEGA references? - talk about a way of promoting MEGA? I have removed all references and replaced with reputable links where possible except the CGMiner 3.7.2 download link - I'll try and find a better one and update later.

Reason for this change is because MEGA is prompting users to download the application/join the site.